Updated standard report flow to add new month filtering page
