+++ server/routes/reporting/monthly-utilisation/monthly-utilisation.controller.js
@@ -1,7 +1,7 @@
 (() => {
   'use strict';
 
-  const { makeManualError } = require('../../../lib/mod-utils');
+  const { makeManualError, generateReportSelectMonths } = require('../../../lib/mod-utils');
   const { dateFilter, capitalizeFully } = require('../../../components/filters');
   const { courtDetailsDAO } = require('../../../objects');
   const { monthlyUtilisationReportsDAO } = require('../../../objects/reports');
@@ -14,8 +14,8 @@
     delete req.session.formFields;
 
     return res.render('reporting/standard-reports/month-select', {
-      months: generateSelectMonths(),
-      reportName: 'Prepare monthly wastage and utilisation',
+      months: generateReportSelectMonths(),
+      reportName: 'Prepare monthly wastage and utilisation report',
       selectMonthLabel: 'Select month to prepare report for',
       processUrl: app.namedRoutes.build('reports.prepare-monthly-utilisation.filter.post'),
       cancelUrl: app.namedRoutes.build('reports.statistics.get'),
@@ -86,7 +86,7 @@
 
       return res.render('reporting/standard-reports/month-select', {
         months,
-        reportName: 'View monthly wastage and utilisation',
+        reportName: 'View monthly wastage and utilisation report',
         selectMonthLabel: 'Select a reporting month',
         showPreviousMonthsCheckbox: true,
         processUrl: app.namedRoutes.build('reports.view-monthly-utilisation.filter.post'),
