+++ server/routes/reporting/standard-report/index.js
@@ -114,8 +114,7 @@
     standardReportRoutes(app, 'jury-summoning-monitor-court');
     require('../jury-summoning-monitor')(app);
     standardReportRoutes(app, 'yield-performance');
-    standardReportRoutes(app, 'all-court-utilisation');
-    require('../digital-summons-received')(app);
+    // require('../digital-summons-received')(app);
     standardReportRoutes(app, 'digital-summons-received');
   };
